Use int for cpus

+++ run/models.py
@@ -212,10 +212,10 @@ class Run(models.Model):
             manager = client.services.create(
                 image='kondor-manager',
                 resources=DockerResources(
-                    cpu_limit=(MANAGER_CPU_LIMIT * (10 ** 9))
+                    cpu_limit=int(MANAGER_CPU_LIMIT * (10 ** 9))
                     if MANAGER_CPU_LIMIT is not None else None,
                     mem_limit=int(MANAGER_MEMORY_LIMIT) if MANAGER_MEMORY_LIMIT is not None else None,
-                    cpu_reservation=(cpu_tot_reserve * (10 ** 9)) if cpu_tot_reserve > 0 else None,
+                    cpu_reservation=int(cpu_tot_reserve * (10 ** 9)) if cpu_tot_reserve > 0 else None,
                     mem_reservation=int(mem_tot_reserve) if mem_tot_reserve > 0 else None,
                 ),
                 env=environment,
